+++ PSGubbins/PSGubbins.psm1
@@ -1,329 +1,4 @@
-﻿# This PSM1 compiled with psake 2020/06/04 22:22:24
-
-# PSGubbins\Source\Functions\Get-CCMLog.ps1
-function Get-CCMLog
-{
-    <#
-    .SYNOPSIS
-        Retrieves entries from Configuration Manager's logs.
-    .DESCRIPTION
-        Retrieves entries from Configuration Manager's log files (defaulting to the last 2000 lines per file) from a local or remote computer and filters using pre-set patterns.
-    .PARAMETER LogName
-        The name of the log to be retrieved.
-    .PARAMETER Path
-        The directory the logs are stored in. UNC path's are assumed for use against remote machine but conversion from local drives is attempted.
-    .PARAMETER ComputerName
-        The computer(s) whose logs will queried.
-    .PARAMETER AllMessages
-        Returns all log entries rather than using the default pattern-matching for filtering.
-    .EXAMPLE
-        Get-CCMLog -ComputerName "lt-12345" -Count 25
-
-        Retrieves the AppEnforce log of lt-12345
-    .EXAMPLE
-        Get-CCMLog -LogName "AppDiscovery"
-
-        Retrieves the AppDiscovery log of the local machine
-    .EXAMPLE
-        Get-CCMLog -LogName "AppIntentEval", "AppDiscovery", "AppEnforce" | Out-GridView
-
-        Retrieves the 'AppIntentEval', 'AppDiscovery' and 'AppEnforce' log entries and outputs to Out-GridView for interactive search and manipulation.
-    .EXAMPLE
-        Get-CCMLog -LogName PolicyAgent, AppDiscovery, AppIntentEval, CAS, ContentTransferManager, DataTransferService, AppEnforce | Out-GridView
-
-        Retrieves logs allowing for the tracing of a deployment from machine policy to app enforcement and outputs to Out-GridView again.
-    .INPUTS
-        String, Integer, Switch
-    .OUTPUTS
-        PSCustomObject
-    #>
-    [CmdletBinding(ConfirmImpact = "Low", SupportsShouldProcess = $True)]
-    [OutputType("PSCustomObject")]
-    param
-    (
-        [Parameter(Position = 0, HelpMessage = "The log(s) to be retrieved")]
-        [alias("Name")]
-        [ValidateSet("AlternateHandler", "AppDiscovery", "AppEnforce", "AppIntentEval", "AssetAdvisor", "CAS", "Ccm32BitLauncher", "CcmCloud",
-            "CcmEval", "CcmEvalTask", "CcmExec", "CcmMessaging", "CcmNotificationAgent", "CcmRepair", "CcmRestart", "CCMSDKProvider",
-            "CcmSqlCE", "CCMVDIProvider", "CertEnrollAgent", "CertificateMaintenance", "CIAgent", "CIDownloader", "CIStateStore",
-            "CIStore", "CITaskMgr", "ClientIDManagerStartup", "ClientLocation", "ClientServicing", "CMBITSManager", "CmRcService",
-            "CoManagementHandler", "ComplRelayAgent", "ContentTransferManager", "DataTransferService", "DCMAgent", "DCMReporting",
-            "DcmWmiProvider", "DdrProvider", "DeltaDownload", "EndpointProtectionAgent", "execmgr", "ExpressionSolver",
-            "ExternalEventAgent", "FileBITS", "FSPStateMessage", "InternetProxy", "InventoryAgent", "InventoryProvider",
-            "LocationServices", "MaintenanceCoordinator", "ManagedProvider", "mtrmgr", "oobmgmt", "PeerDPAgent", "PolicyAgent",
-            "PolicyAgentProvider", "PolicyEvaluator", "pwrmgmt", "PwrProvider", "RebootCoordinator", "ScanAgent", "Scheduler",
-            "ServiceWindowManager", "SettingsAgent", "setuppolicyevaluator", "smscliui", "SoftwareCatalogUpdateEndpoint",
-            "SoftwareCenterSystemTasks", "SrcUpdateMgr", "StateMessage", "StateMessageProvider", "StatusAgent", "SWMTRReportGen",
-            "UpdatesDeployment", "UpdatesHandler", "UpdatesStore", "UpdateTrustedSites", "UserAffinity", "UserAffinityProvider",
-            "VirtualApp", "wedmtrace", "WindowsAnalytics", "WUAHandler")]
-        [string[]] $LogName = "AppEnforce",
-
-        [Parameter(Position = 1, HelpMessage = "The path to the directory containing the logs")]
-        [string] $Path = "C:\Windows\CCM\Logs",
-
-        [Parameter(Position = 3, ValueFromPipeline = $True, ValueFromPipelineByPropertyName = $True, HelpMessage = "The computer whose logs will be parsed")]
-        [alias("PSComputerName", "__SERVER", "CN", "IPAddress")]
-        [string[]] $ComputerName = "localhost",
-
-        [Parameter(HelpMessage = "Returns all log entries rather than using the default pattern-matching for filtering")]
-        [alias("NoPattern")]
-        [switch] $AllMessages
-    )
-
-    BEGIN
-    {
-        if ($PSBoundParameters.ContainsKey("Debug"))
-        {
-            $DebugPreference = "Continue"
-        }
-
-        $PatternContents = if ($AllMessages)
-        {
-            "(.+)"
-        }
-        else
-        {
-            "(a?)sync(hronous?)"
-            "(Prepared|Executing) command line"
-            "App enforcement completed"
-            "Application not discovered"
-            "authority"
-            "cache"
-            "complete(d?)"
-            "criteria"
-            "detect(ed?)"
-            "exclude"
-            "exit code"
-            "fail(ure|s|ed)"
-            "include"
-            "Install enforcement"
-            "missing"
-            "Performing detection of app deployment type"
-            "policy"
-            "Post install behavior"
-            "Prepared working directory"
-            "registration"
-            "releas(e|ed|ing)"
-            "request(ed|ing)?"
-            "revok(e|ed|ing)"
-            "S-1-5"
-            "search"
-            "set"
-            "size"
-            "source"
-            "state"
-            "status"
-            "succe(eded|ss)"
-            "tombstone"
-            "update(d?)"
-            "user logged"
-            "Waiting for process"
-        }
-        # Convert the array to a regex pattern
-        $Pattern = $PatternContents -join "|"
-    }
-
-    PROCESS
-    {
-        forEach ($Computer in $ComputerName)
-        {
-            if (($Computer -like "localhost") -or ($Computer -like "127.0.0.1"))
-            {
-                Write-Verbose -Message "Converting localhost to hostname"
-                $Computer = [Environment]::MachineName
-            }
-            Write-Verbose -Message "Processing '$Computer'"
-
-            try
-            {
-                $LogRoot = $Null
-                if ($Null -ne $Path)
-                {
-                    $JoinSplat = @{
-                        Path        = "\\$Computer\"
-                        ChildPath   = ($Path -replace ":", "$")
-                        ErrorAction = "Stop"
-                    }
-
-                    $UNCPath = Join-Path @JoinSplat
-                    if (Test-Path -Path $UNCPath -ErrorAction "Stop")
-                    {
-                        $LogRoot = Resolve-Path -Path $UNCPath -ErrorAction "Stop"
-                    }
-                }
-                else
-                {
-                    $LogRoot = Resolve-Path "\\$Computer\C$\Windows\CCM\Logs\" -ErrorAction "Stop"
-                }
-
-                if ($LogRoot -match "^Microsoft.Powershell")
-                {
-                    $LogRoot = $LogRoot -Split "::" | Select-Object -Last 1
-                }
-            }
-            catch
-            {
-                Write-Warning -Message "Problems were encountered resolving '$LogRoot'"
-                $PSCmdlet.ThrowTerminatingError($_)
-            }
-
-            if (-not (Test-Path -Path $LogRoot))
-            {
-                Write-Warning -Message "Unable to resolve/access '$LogRoot'"
-                Continue
-            }
-
-            Write-Verbose -Message "Testing for connectivity to '$Computer'"
-            if (($Computer -eq [Environment]::MachineName) -or
-                (Test-Connection -ComputerName $Computer -Quiet -Count 2))
-            {
-                forEach ($Log in $LogName)
-                {
-                    if ($PSCmdlet.ShouldProcess($LogRoot, "Retrieve '$Log' log entries"))
-                    {
-                        try
-                        {
-                            Write-Verbose -Message "Locating '$Log' log(s)."
-                            $LogSearchSplat = $Null
-                            $LogSearchSplat = @{
-                                Path   = $LogRoot
-                                Filter = "$Log*.log"
-                                File   = $True
-                            }
-
-                            $LogPaths = $Null
-                            $LogPaths = @(Get-ChildItem @LogSearchSplat  | Select-Object -ExpandProperty "FullName")
-                            Write-Verbose -Message "'$($LogPaths.Count)' '$Log' logs found."
-                        }
-                        catch
-                        {
-                            Write-Warning -Message "Problems were encountered '$Log' logs from '$LogRoot'"
-                            $PSCmdlet.ThrowTerminatingError($_)
-                        }
-
-                        forEach ($LogPath in $LogPaths)
-                        {
-                            if (Test-Path -Path $LogPath -ErrorAction "Stop")
-                            {
-                                try
-                                {
-                                    $Parameters = @{
-                                        Path        = $LogPath
-                                        Tail        = 2000     # I'm reluctant to read the entirety of the files by default and this seems likely to read most logs.
-                                        ErrorAction = "Stop"
-                                    }
-
-                                    Write-Verbose -Message "Reading log ($LogPath)."
-                                    $LogContents = Get-Content @Parameters |
-                                        Where-Object { $_ -match $Pattern }
-                                }
-                                catch [System.UnauthorizedAccessException]
-                                {
-                                    Write-Warning -Message "Unable due to retrieve details to an 'Unauthorized Access' exception. Ensure you have the required permissions."
-                                    $PSCmdlet.ThrowTerminatingError($_)
-                                }
-                                catch
-                                {
-                                    $PSCmdlet.ThrowTerminatingError($_)
-                                }
-
-                                forEach ($LogEntry in $LogContents)
-                                {
-                                    Write-Debug -Message "Processing log entry: '$LogEntry'"
-                                    try
-                                    {
-                                        Write-Debug -Message "Processing message property."
-                                        [string] $Message = ($LogEntry | Select-String -Pattern "\[LOG\[((.| )+)\]LOG\]").Matches.Value
-                                        # Identifies the message sections of each line using the [LOG[] tags and removes them for us.
-                                        if ($Log -like "AppIntentEval")
-                                        {
-                                            Write-Debug -Message "Parsing AppIntentEval message."
-                                            $Message = $Message -replace ":- |, ", "`n"
-                                            # Lazy reformatting to key:value statements instead of in a line.
-                                        }
-                                        $Message = ($Message -replace "\[LOG\[|\]LOG\]").Trim()
-                                    }
-                                    catch
-                                    {
-                                        $PSCmdlet.ThrowTerminatingError($_)
-                                    }
-
-                                    if (-not ($Message))
-                                    {
-                                        Write-Verbose -Message "Unable to read blank message - skipping to next."
-                                        Continue
-                                    }
-
-                                    try
-                                    {
-                                        Write-Debug -Message "Processing metadata block."
-                                        [string] $Metadata = ((($LogEntry | Select-String -Pattern "<time((.| )+)>").Matches.Value -replace "<|>") -split " ")
-                                        # Identifies and isolates the metadata section
-                                        # Includes the time and date which we want, but also other entries which we'll need to get rid of.
-                                    }
-                                    catch
-                                    {
-                                        $PSCmdlet.ThrowTerminatingError($_)
-                                    }
-
-
-                                    try
-                                    {
-                                        Write-Debug -Message "Processing TimeStub block."
-                                        [string] $TimeStub = ((($Metadata -split " ")[0] -replace 'time|=|"') -split "\.")[0]
-                                        # To find the time, we remove 'time', '=', and '"'
-                                        # Split the remainder in two based on '.' and keep the first part.
-                                        # This is a bit awkward but the casting is tricky without the split.
-                                    }
-                                    catch
-                                    {
-                                        $PSCmdlet.ThrowTerminatingError($_)
-                                    }
-
-                                    try
-                                    {
-                                        Write-Debug -Message "Processing DateStub block."
-                                        [string] $DateStub = ((($Metadata -split " ")[1] -replace 'date|=|"'))
-                                        # Finding the date is similar but simpler.
-                                        # We only need to remove 'date', '=', and '"'.
-                                    }
-                                    catch
-                                    {
-                                        $PSCmdlet.ThrowTerminatingError($_)
-                                    }
-
-
-                                    try
-                                    {
-                                        Write-Debug -Message "Generating timestamp."
-                                        [datetime] $TimeStamp = "$DateStub $TimeStub"
-                                        # At the end we add the two stubs together, and cast them as a [datetime] object
-                                    }
-                                    catch
-                                    {
-                                        $PSCmdlet.ThrowTerminatingError($_)
-                                    }
-
-                                    [PSCustomObject]@{
-                                        ComputerName = $Computer
-                                        Source       = $Log
-                                        Timestamp    = $TimeStamp
-                                        Message      = $Message
-                                        Path         = $LogPath
-                                    }
-                                }
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    END
-    {
-    }
-}
+# This PSM1 compiled with psake 2020/11/01 20:28:21
 
 # PSGubbins\Source\Functions\Get-Cpl.ps1
 function Get-Cpl
@@ -553,24 +228,36 @@ function New-Password
         [int] $MaximumLength = $Null,
 
         [Parameter(Mandatory = $False, HelpMessage = "The first group of words to be used when generating the password - adjectives by default")]
-        [string[]] $FirstDictionary = @("Angry", "Annoying", "Awesome", "Basic", "Bitter", "Clever", "Correct",
-            "Divine", "Enraged", "Fast", "Fat", "Great", "Guilty", "Handy", "Huge", "Intelligent", "Irritated",
-            "Just", "Massive", "Moral", "Morose", "Psyched", "Rapid", "Resourceful", "Righteous", "Sad", "Scary",
-            "Sinful", "Smart", "Strong", "Tiny", "Upright", "Vengeful", "Vexed", "Violent", "Wrathful"),
+        [string[]] $FirstDictionary = @('Adorable', 'Adventurous', 'Aggressive', 'Agreeable', 'Alert', 'Alive', 'Amused', 'Angry', 'Annoyed', 'Annoying', 'Anxious',
+            'Arrogant', 'Ashamed', 'Attractive', 'Average', 'Awesome', 'Awful', 'Bad', 'Basic', 'Beautiful', 'Better', 'Bewildered', 'Bitter', 'Black', 'Bloody',
+            'Blue', 'Blue-eyed', 'Blushing', 'Bored', 'Brainy', 'Brave', 'Breakable', 'Bright', 'Busy', 'Calm', 'Careful', 'Cautious', 'Charming', 'Cheerful',
+            'Clean', 'Clear', 'Clever', 'Cloudy', 'Clumsy', 'Colorful', 'Combative', 'Comfortable', 'Concerned', 'Condemned', 'Confused', 'Cooperative', 'Correct',
+            'Courageous', 'Crazy', 'Creepy', 'Crowded', 'Cruel', 'Curious', 'Cute', 'Dangerous', 'Dark', 'Dead', 'Defeated', 'Defiant', 'Delightful', 'Depressed',
+            'Determined', 'Different', 'Difficult', 'Disgusted', 'Distinct', 'Disturbed', 'Divine', 'Dizzy', 'Doubtful', 'Drab', 'Dull', 'Eager', 'Easy', 'Elated',
+            'Elegant', 'Embarrassed', 'Enchanting', 'Encouraging', 'Energetic', 'Enraged', 'Enthusiastic', 'Envious', 'Evil', 'Excited', 'Expensive', 'Exuberant',
+            'Fair', 'Faithful', 'Famous', 'Fancy', 'Fantastic', 'Fast', 'Fat', 'Fierce', 'Filthy', 'Fine', 'Foolish', 'Fragile', 'Frail', 'Frantic', 'Friendly',
+            'Frightened', 'Funny', 'Gentle', 'Gifted', 'Glamorous', 'Gleaming', 'Glorious', 'Good', 'Gorgeous', 'Graceful', 'Great', 'Grieving', 'Grotesque',
+            'Grumpy', 'Guilty', 'Handsome', 'Handy', 'Happy', 'Healthy', 'Helpful', 'Helpless', 'Hilarious', 'Homeless', 'Homely', 'Horrible', 'Huge', 'Hungry',
+            'Hurt', 'Ill', 'Important', 'Impossible', 'Inexpensive', 'Innocent', 'Inquisitive', 'Intelligent', 'Irritated', 'Itchy', 'Jealous', 'Jittery', 'Jolly',
+            'Joyous', 'Just', 'Kind', 'Lazy', 'Light', 'Lively', 'Lonely', 'Long', 'Lovely', 'Lucky', 'Magnificent', 'Massive', 'Misty', 'Modern', 'Moral', 'Morose',
+            'Motionless', 'Muddy', 'Mushy', 'Mysterious', 'Nasty', 'Naughty', 'Nervous', 'Nice', 'Nutty', 'Obedient', 'Obnoxious', 'Odd', 'Open', 'Outrageous',
+            'Outstanding', 'Panicky', 'Perfect', 'Plain', 'Pleasant', 'Poised', 'Poor', 'Powerful', 'Precious', 'Prickly', 'Proud', 'Psyched', 'Putrid', 'Puzzled',
+            'Quaint', 'Rapid', 'Real', 'Relieved', 'Repulsive', 'Resourceful', 'Rich', 'Righteous', 'Sad', 'Scary', 'Selfish', 'Shiny', 'Shy', 'Silly', 'Sinful',
+            'Sleepy', 'Smart', 'Smiling', 'Smoggy', 'Sore', 'Sparkling', 'Splendid', 'Spotless', 'Stormy', 'Strange', 'Strong', 'Stupid', 'Successful', 'Super',
+            'Talented', 'Tame', 'Tasty', 'Tender', 'Tense', 'Terrible', 'Thankful', 'Thoughtful', 'Thoughtless', 'Tiny', 'Tired', 'Tough', 'Troubled', 'Ugliest',
+            'Ugly', 'Uninterested', 'Unsightly', 'Unusual', 'Upright', 'Upset', 'Uptight', 'Vast', 'Vengeful', 'Vexed', 'Victorious', 'Violent', 'Vivacious',
+            'Wandering', 'Weary', 'Wicked', 'Wild', 'Witty','Worried', 'Worrisome', 'Wrathful', 'Wrong', 'Zany', 'Zealous'),
 
         [Parameter(Mandatory = $False, HelpMessage = "The second group of words to be used when generating the password - animal names by default")]
-        [string[]] $SecondDictionary = @("Alligators", "Apes", "Baboons", "Badgers", "Beavers", "Cats",
-            "Chipmunks", "Coyotes", "Crocodiles", "Dinsoaurs", "Dogs", "Dolphins", "Ducks", "Ferrets", "Fish",
-            "Fleas", "Foxes", "Geese", "Giraffes", "Hares", "Honey Badgers", "Horses", "Leopards", "Lions",
-            "Meerkats", "Mice", "Monkeys", "Rabbits", "Raccoons", "Rats", "Sharks", "Squirrels", "Swans",
-            "Tigers", "Tortoises", "Turtles", "Weasels", "Wolves", "Zebras"),
+        [string[]] $SecondDictionary = @('Alligators', 'Apes', 'Baboons', 'Badgers', 'Beavers', 'Cats', 'Cheetahs', 'Chipmunks', 'Coyotes', 'Crocodiles', 'Crocs',
+            'Dinosaurs', 'Dogs', 'Dolphins', 'Ducks', 'Ferrets', 'Fish', 'Fleas', 'Foxes', 'Gators', 'Geese', 'Giraffes', 'Goats', 'Hares', 'Honey Badgers', 'Horses',
+            'Leopards', 'Lions', 'Meerkats', 'Mice', 'Monkeys', 'Moose', 'Pandas', 'Penguins', 'Rabbits', 'Raccoons', 'Rats', 'Reindeer', 'Sharks', 'Sheep',
+            'Squirrels', 'Swans', 'Tigers', 'Tortoises', 'Turtles', 'Weasels', 'Wolves', 'Zebras'),
 
         [Parameter(Mandatory = $False, HelpMessage = "The third group of words to be used when generating the password - modifiers by default")]
-        [string[]] $ExtraDictionary = @("Abhorrently", "Abnormally", "Absurdly", "Acceptably", "Accordingly",
-            "Adorably", "Amazingly", "Artfully", "Creatively", "Extremely", "Incredibly", "Infinitely", "Mad",
-            "Moderately", "Particularly", "Pleasingly", "Proper", "Really", "Reasonably", "Somewhat",
-            "Strikingly", "Sufficiently", "Super", "Supremely", "Totally", "Tremendously", "Truly", "Very",
-            "Well", "Wicked")
+        [string[]] $ExtraDictionary = @("Abhorrently", "Abnormally", "Absurdly", "Acceptably", "Accordingly", "Adorably", "Amazingly", "Artfully", "Creatively",
+            "Extremely", "Incredibly", "Infinitely", "Mad", "Moderately", "Particularly", "Pleasingly", "Proper", "Really", "Reasonably", "Somewhat", "Strikingly",
+            "Sufficiently", "Super", "Supremely", "Totally", "Tremendously", "Truly", "Very", "Well", "Wicked")
     )
 
     BEGIN